+++ src/js/relevant.js
@@ -191,6 +191,9 @@ export default {
             this.form.calc.update( {
                 relevantPath: path
             } );
+            this.form.itemset.update( {
+                relevantPath: path
+            } );
             // Update outputs that are children of branch
             // TODO this re-evaluates all outputs in the form which is not efficient!
             this.form.output.update();