+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 --------------------
 ##### Fixed
 - Loading error in Pages mode when a page has no label.
+- Itemsets not populating if inside an irrelevant group upon load.
 
 [5.0.1] - 2018-12-20
 --------------------